requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
ollama
pypdf
faiss-cpu
numpy
langchain_text_splitters
PyMuPDF
